restart1er winform, correctif service

+++ Whyne.DAL/Item.cs
@@ -11,7 +11,7 @@ namespace Whyne.DAL
         [Required] public string? Icon { get; set; }
         [MaxLength(150)] [Required] public string? Title { get; set; }
         [Required] public string? Description { get; set; }
-        [Required] public int? DegreeAlcoolo { get; set; }
+        [Required] public float? DegreeAlcoolo { get; set; }
         [MaxLength(150)] [Required] public string? Familly { get; set; }
         [Required] public int? Quantity { get; set; }
         [MaxLength(150)] [Required] public string? Provider { get; set; }
